requirements.txt
Cython==0.29.24
imageio==2.9.0
opencv-contrib-python==4.1.2.30
Pillow==8.3.2
pycocotools==2.0.2
python-dateutil==2.8.2
requests==2.26.0
torch
torchaudio
torchvision
tqdm==4.62.3
urllib3==1.26.6
